refactor: move prod env detection to init stage
